clip_l1/requirements.txt
numpy
opencv-python
torch
requests
transformers
Pillow